Update chapter 8 code with new stdlib library Update chapter 9 and code with removed PHYSICAL_TIMER definition and stdlib code
