+++ tutorial/08-generalization/code/libraries/CMakeLists.txt
@@ -2,3 +2,4 @@ message(STATUS "\n**************************************************************
 message(STATUS "\n## In directory: ${CMAKE_CURRENT_SOURCE_DIR}")
 
 add_subdirectory(baremetal)
+add_subdirectory(stdlib)
