+++ tutorial/08-generalization/code/libraries/baremetal/include/baremetal/IMemoryAccess.h
@@ -39,7 +39,7 @@
 
 #pragma once
 
-#include <baremetal/Types.h>
+#include <stdlib/Types.h>
 
 /// @file
 /// Abstract memory access interface
