+++ tutorial/CMakeLists.txt
@@ -7,3 +7,4 @@ add_subdirectory(06-console-uart1)
 add_subdirectory(07-improving-startup-static-initialization)
 add_subdirectory(08-generalization)
 add_subdirectory(09-timer)
+#add_subdirectory(10-mailbox)
